onlineportal/requirements.txt
Django==3.0.5
django-ranged-response
django-simple-captcha==0.5.12
gunicorn==20.0.4
Pillow==7.1.2
psycopg2-binary==2.8.5
six==1.14.0
pip==20.1.0
django-crispy_forms==1.9.0